fix(cli): stop local dev stack on exit (#2914)

+++ apps/openci_cli/lib/src/commands/dev/start_orchard_worker.dart
@@ -88,7 +88,8 @@ class OrchardWorker {
   }
 
   void _handleSignal(ProcessSignal signal) {
-    _stopSignal ??= signal;
+    if (_stopRequested) return;
+    _stopSignal = signal;
     _stopRequested = true;
     _process.kill(ProcessSignal.sigint);
   }
